fixed: fixed file layout and some small bugs
