--- src/ImagesUels.js
@@ -1,29 +1,0 @@
-// store all your watch images on imgur
-// export the links with the DIRECT LINK choice
-// so that you get the .png extension
-// dump them in an array
-// export them
-// import them elsewhere for use
-// or, iterate over these values when you seed
-
-export const imageUrls = [
-  "https://i.imgur.com/sm4i82J.png",
-  "https://i.imgur.com/ed1G33y.png",
-  "https://i.imgur.com/9M20e1b.png",
-  "https://i.imgur.com/T6UediO.png",
-  "https://i.imgur.com/cvqE34m.png?1",
-  "https://i.imgur.com/HfEls2w.png",
-  "https://i.imgur.com/jkSzrTz.png",
-  "https://i.imgur.com/bSICAJ1.png",
-  "https://i.imgur.com/I3h729Y.png",
-  "https://i.imgur.com/5gpC9Q9.png",
-  "https://i.imgur.com/yavFkOG.png",
-  "https://i.imgur.com/TyGAIFH.png",
-  "https://i.imgur.com/0fax8zt.png",
-  "https://i.imgur.com/lS2CtBm.png",
-];
-
-/* 
-   in the react component,
-   just do imageUrls.map((url, idx) => <img src={url} key={idx} alt={'watch-image'} />)
-*/
