+++ src/components/index.js
@@ -8,5 +8,6 @@ export { default as Footer } from "./Footer";
 export { default as Nav } from "./Nav";
 export { default as ProductView } from "./ProductView";
 export { default as ProductRows } from "./ProductRows";
-export { default as SignupOrLogin } from "./SignupOrLogin";
+export { default as Signup } from "./Signup";
+export { default as Login } from "./Login"
 export { default as Me } from "./Me";
