+++ src/components/styles/Login.css
@@ -1,0 +1,62 @@
+h1 {
+  color: #fff;
+  text-shadow: 0 0 10px rgb(0 0 0 / 30%);
+  letter-spacing: 1.1px;
+  text-align: center;
+  margin-bottom: 1.2em;
+}
+
+.login_container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 30px;
+}
+
+form {
+  display: flex;
+  height: 180px;
+  /* width: 250px; */
+  /* flex-direction: column; */
+  /* justify-content: space-evenly; */
+  /* align-items: flex-start; */
+  margin: 1em;
+}
+
+input {
+  height: 2.5em;
+  width: 100%;
+  color: #fff;
+  background-color: rgba(0, 0, 0, 0.3);
+  border-radius: 0.3em;
+  border: none;
+  outline: none;
+  padding: 5px;
+  font-size: 13px;
+  text-shadow: 1px 1px 1px rgba(20, 40, 20, 0.3);
+  box-shadow: inset 0 -5px 45px rgba(80, 40, 20, 0.2),
+    0 1px 1px rgba(255, 255, 255, 0.2);
+}
+
+input:-webkit-autofill,
+input:-webkit-autofill:hover,
+input:-webkit-autofill:focus,
+input:-webkit-autofill:active {
+  -webkit-text-fill-color: #fff !important;
+
+  transition: background-color 5000s ease-in-out 0s;
+}
+
+.loginBtn {
+  height: 2.5em;
+  width: 100%;
+  color: #fff;
+  cursor: pointer;
+  border: 0.1em solid #141414;
+  border-radius: 0.3em;
+  background-color: #2b4af1;
+  background-image: linear-gradient(0deg, #2b4af1 0%, #749bff 100%);
+  text-shadow: 0 1.1px 2px rgba(0, 0, 0, 0.336);
+  box-shadow: inset 0 0.05em 0.1em rgba(32, 31, 31, 0.671);
+}
